Rename mutation file names
