+++ src/matches/models/match.js
@@ -1,6 +1,6 @@
 import getMatches from '../queries/getMatches';
-import matchRequest from '../mutations/match-request';
-import updateMatch from '../mutations/update-match';
+import matchRequest from '../mutations/matchRequest';
+import updateMatch from '../mutations/updateMatch';
 
 module.exports = {
   getMatches: getMatches,
