+++ src/profiles/models/profile.js
@@ -1,6 +1,6 @@
 import getProfile from '../queries/getProfile';
-import createProfile from '../mutations/create-profile';
-import updateProfile from '../mutations/update-profile';
+import createProfile from '../mutations/createProfile';
+import updateProfile from '../mutations/updateProfile';
 
 module.exports = {
   getProfile: getProfile,
