+++ src/profiles/mutations/createProfile.js
@@ -5,7 +5,7 @@ import uuidv4 from 'uuid/v4';
 
 const CREATED_PROFILE = 'CREATED_PROFILE';
 
-const insertCreatedProfileEvent = ({ id, userId, bio, streamId }, trx) => {
+const insertCreatedProfileEvent = ({ profileId, userId, bio, streamId }, trx) => {
   const eventId = uuidv4();
   return pg('event_sourcing.event_store')
     .transacting(trx)
@@ -13,7 +13,7 @@ const insertCreatedProfileEvent = ({ id, userId, bio, streamId }, trx) => {
       id: eventId,
       event_type: CREATED_PROFILE,
       payload: {
-        id,
+        id: profileId,
         user_id: userId,
         bio,
       },
@@ -24,7 +24,8 @@ const insertCreatedProfileEvent = ({ id, userId, bio, streamId }, trx) => {
     });
 }
 
-const insertProfile = ({ id, bio, userId }, trx) => {
+const insertProfile = ({ userId, bio }, trx) => {
+  const id = uuidv4();
   return pg('core.users_profile')
     .transacting(trx)
     .insert({
@@ -37,21 +38,23 @@ const insertProfile = ({ id, bio, userId }, trx) => {
     .then(humps.camelizeKeys)
 };
 
-const createProfile = ({ userId, bio }) => {
-  const id = uuidv4();
+const createProfile = ({ bio }, { user, isAuthenticated }) => {
+  const userId = user.id;
   let profile;
+  if (!isAuthenticated) {
+    throw new Error(`User is not authenticated`);
+  }
   return pg.transaction(trx => {
     return insertProfile({
-      id,
       userId,
       bio
     }, trx)
       .then(result => profile = result)
       .then(() => insertCreatedProfileEvent({
-        id,
+        profileId: profile.id,
         userId,
         bio,
-        streamId: id
+        streamId: profile.id
       }, trx))
       .then(trx.commit)
       .catch(trx.rollback)
