+++ src/users/models/user.js
@@ -1,9 +1,9 @@
 import getUsers from '../queries/getUsers';
 import getUser from '../queries/getUser';
 import getRandomUsers from '../queries/getRandomUsers';
-import loginUser from '../mutations/login-user';
-import createUser from '../mutations/create-user';
-import updateUser from '../mutations/update-user';
+import loginUser from '../mutations/loginUser';
+import createUser from '../mutations/createUser';
+import updateUser from '../mutations/updateUser';
 
 module.exports = {
   getUsers: getUsers,
